# Optimise double pipeline