`optimise-no-content` branch:
```
> time curl  -s 'localhost:8080/api/preprints-no-content?per-page=100&page=1'  > /dev/null

real	0m0.648s
user	0m0.006s
sys	0m0.011s

[09:20:21] 650ms 0  ✔️
> time curl  -s 'localhost:8080/api/preprints-no-content?per-page=100&page=2'  > /dev/null

real	0m0.419s
user	0m0.006s
sys	0m0.012s

[09:20:25] 417ms 0  ✔️
> time curl  -s 'localhost:8080/api/preprints-no-content?per-page=100&page=3'  > /dev/null

real	0m0.483s
user	0m0.005s
sys	0m0.009s

[09:20:27] 492ms 0  ✔️
```

This branch:
```
> time curl  -s 'localhost:8080/api/preprints-no-content?per-page=100&page=1'  > /dev/null

real	0m0.151s
user	0m0.006s
sys	0m0.013s

[09:21:27] 140ms 0  ✔️
> time curl  -s 'localhost:8080/api/preprints-no-content?per-page=100&page=2'  > /dev/null

real	0m0.062s
user	0m0.006s
sys	0m0.011s

[09:21:30] 64ms 0  ✔️
> time curl  -s 'localhost:8080/api/preprints-no-content?per-page=100&page=3'  > /dev/null

real	0m0.065s
user	0m0.006s
sys	0m0.012s

[09:21:32] 65ms 0  ✔️
```